The plot thickens.
This runs at home on 2.3.3 and crashes on the spinner editor
opening during onMeasure

+++ src/com/eyebrowssoftware/bptrackerfree/activity/BPRecordEditor.java
@@ -124,18 +124,17 @@ public class BPRecordEditor extends BPRecordEditorBase implements OnItemSelected
     }
 
     @Override
-    public void onItemSelected(AdapterView<?> parent, View view, int pos,
-            long id) {
+    public void onItemSelected(AdapterView<?> parent, View view, int pos, long id) {
         Spinner sp = (Spinner) parent;
         if (sp.equals(mSpinners[SYS_IDX])) {
-            int systolic = (Integer) ((RangeAdapter) mSpinners[SYS_IDX].getAdapter()).getItem(pos);
+            int systolic = ((RangeAdapter) mSpinners[SYS_IDX].getAdapter()).getItem(pos);
             int diastolic = (Integer) mSpinners[DIA_IDX].getSelectedItem();
             if ((systolic - diastolic) < BPTrackerFree.MIN_RANGE) {
                 Toast.makeText(this, R.string.check_values, Toast.LENGTH_SHORT).show();
             }
         } else if (sp.equals(mSpinners[DIA_IDX])) {
             int systolic = (Integer) mSpinners[SYS_IDX].getSelectedItem();
-            int diastolic = (Integer) ((RangeAdapter) mSpinners[DIA_IDX].getAdapter()).getItem(pos);
+            int diastolic = ((RangeAdapter) mSpinners[DIA_IDX].getAdapter()).getItem(pos);
             if ((systolic - diastolic) < BPTrackerFree.MIN_RANGE) {
                 Toast.makeText(this, R.string.check_values, Toast.LENGTH_SHORT).show();
             }
